backend/requirements.txt
fastapi
uvicorn
firebase-admin
google-genai
google-adk
opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk
pydantic
websockets
python-dotenv
requests
aiohttp